+++ src/vec4.hpp
@@ -7,24 +7,24 @@ class mat4;
 
 class vec4 {
 public:
-    double x, y, z, w;
+    float x, y, z, w;
     vec4(): x(0), y(0), z(0), w(0) {}
-    vec4(double a, double b, double c): x(a), y(b), z(c), w(0) {}
-    vec4(double a, double b, double c, double d): x(a), y(b), z(c), w(d) {}
+    vec4(float a, float b, float c): x(a), y(b), z(c), w(0) {}
+    vec4(float a, float b, float c, float d): x(a), y(b), z(c), w(d) {}
 
     vec4 normalize();
     vec4 cross(vec4 &v);
-    double dot(vec4 &v);
-    double length();
+    float dot(vec4 &v);
+    float length();
 
     vec4 operator + (const vec4 & v);
     vec4 & operator += (const vec4 & v);
     vec4 operator - (const vec4 & v);
     vec4 & operator -= (const vec4 & v);
-    vec4 operator * (double c);
-    vec4 & operator *= (double c);
-    vec4 operator / (double c);
-    vec4 & operator /= (double c);
+    vec4 operator * (float c);
+    vec4 & operator *= (float c);
+    vec4 operator / (float c);
+    vec4 & operator /= (float c);
 };
 
 inline std::ostream& operator << (std::ostream &os, const vec4 &t) {
